Plugins: moving tools to plugins folder
