+++ pydatview/plugins/tool_logdec.py
@@ -1,0 +1,39 @@
+import wx
+from pydatview.plugins.base_plugin import GUIToolPanel
+from pydatview.tools.damping import logDecFromDecay
+
+# --------------------------------------------------------------------------------}
+# --- Log Dec
+# --------------------------------------------------------------------------------{
+class LogDecToolPanel(GUIToolPanel):
+    def __init__(self, parent):
+        super(LogDecToolPanel,self).__init__(parent)
+        btClose = self.getBtBitmap(self,'Close'  ,'close'  ,self.destroy  )
+        btComp  = self.getBtBitmap(self,'Compute','compute',self.onCompute)
+        self.lb = wx.StaticText( self, -1, '                     ')
+        self.sizer = wx.BoxSizer(wx.HORIZONTAL)
+        self.sizer.Add(btClose   ,0, flag = wx.LEFT|wx.CENTER,border = 1)
+        self.sizer.Add(btComp    ,0, flag = wx.LEFT|wx.CENTER,border = 5)
+        self.sizer.Add(self.lb   ,0, flag = wx.LEFT|wx.CENTER,border = 5)
+        self.SetSizer(self.sizer)
+
+    def onCompute(self,event=None):
+        if len(self.parent.plotData)!=1:
+            Error(self,'Log Dec tool only works with a single plot.')
+            return
+        pd =self.parent.plotData[0]
+        try:
+            logdec,DampingRatio,T,fn,fd,IPos,INeg,epos,eneg=logDecFromDecay(pd.y,pd.x)
+            lab='LogDec.: {:.4f} - Damping ratio: {:.4f} - F_n: {:.4f} - F_d: {:.4f} - T:{:.3f}'.format(logdec,DampingRatio,fn,fd,T)
+            self.lb.SetLabel(lab)
+            self.sizer.Layout()
+            ax=self.parent.fig.axes[0]
+            ax.plot(pd.x[IPos],pd.y[IPos],'o')
+            ax.plot(pd.x[INeg],pd.y[INeg],'o')
+            ax.plot(pd.x ,epos,'k--')
+            ax.plot(pd.x ,eneg,'k--')
+            self.parent.canvas.draw()
+        except:
+            self.lb.SetLabel('Failed. The signal needs to look like the decay of a first order system.')
+        #self.parent.load_and_draw(); # DATA HAS CHANGED
+
